board-rskrza1: Fix LCD display shift for RGB565
The start address for the frame buffer must be PAGE aligned. Otherwise when
you mmap /dev/fb0 in user space, mmap will return a PAGE aligned address
that will be different than what the VDC5 was programmed with. Then where
the app think the image starts and where the VDC5 thinks the image starts
will be different, hence the display shift.