rskrza1: Add continugous memory declaration for CEU in XIP Mode
The CEU requires a large amount of contiguous memory. Unfortunately,
fragmentation occurs very quickly in such a small space, so it breaks
the driver.

To get around this, we reserve a small amount of space at the end of
memory using the mem= command line option, then manually allocate
that memory outside of the kernel memory management.