spi-sh-spibsc: Add support for Micron N25Q00AA SPI Flash
The existing driver was hard-coded for the Spansion flash on the RSK board.
This separates out the Spansion specific code and allows for other SPI
flash device to be supported.
Specifically, the Micron N25Q00AA was added and tested. However, note that
at this time, you can only do read-only access with the N25Q00AA (erase and
program commands will not work)