spi-sh-spibsc: Bug fix and allow to be used with XIP_KERNEL
Bug fix: If you didn't add your SPI device to the to dummy_cycle_table[],
dev_param would not be set and the kernel would crash on boot.

XIP: While you cannot enable the SPIBSC driver on a QSPI channel that you
are using for XIP execution, you can use the SPIBSC driver on the other
spare QSPI channel safely.