rskrza1: Remove unneeded MMU mappings
These mappings are not needed. For early BSP releases, they were needed
because some drivers tried to talk directly to the registers using their
physical addresses. However, now the mappings are done by the Device Tree
code, so we can delete this code (it is messy to have io mappings for no
reason)