vdc5fb: Allow virtual LCD sizes on a fixed resolution panel
It is possible to connect the VDC5 to a LCD of a fixed resolution, but
then make the software believe that is has a smaller sized LCD connected.
For example, you can use the RSK 800x480 LCD but make Linux think it has
a QVGA screen attached (that uses less RAM for frame buffer).

To do this, panel_pixel_xres is set to the physical LCD screen size,
but then you set fb_videomode.xres and layers[2].xres to something smaller
like 640. You can also set the Y values as well.