streamit: remove PERF_EVENTS=y from defconfig
The reason is that when this is enabled, the ARM hardware breakpoint driver
(arch/arm/kernel/hw_breakpoint.c) is also enabled and for some reason, the
init function in that driver crashed the kernel boot. This might be caused
by some issue with the JTAG connection circuit design of the board.